Implement basic functionality
This implementation allows to create a püluigin with minimal
boilerplate. Everything in the `lib` folder should go into a separate
library that is then required by a plugin developer so that they can
simply create the file `src/Plugin.php` and that's it. Everything else
is then handled by the library-package